+++ wb_animation_action/utils/utils.py
@@ -1,0 +1,32 @@
+#!/usr/bin/env python3
+#
+# Copyright 1996-2020 Cyberbotics Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import shutil
+
+
+def is_debug():
+    return 'DEBUG' in os.environ and os.environ['DEBUG']
+
+
+def remove_anything(path):
+    """Deletes whatever the `path` value is, folder or file."""
+    if os.path.isfile(path) or os.path.islink(path):
+        os.unlink(path)
+    elif os.path.isdir(path):
+        shutil.rmtree(path)
+    else:
+        print(f'`{path}` is not folder or file!')
